backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.40
pymysql==1.1.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.20
pydantic-settings==2.8.1
pandas==2.2.3
numpy==2.2.4
scikit-learn==1.6.1
networkx==3.4.2
joblib==1.4.2
Pillow==10.4.0
PyPDF2==3.0.1
openpyxl==3.1.5
pytesseract==0.3.13
